Code Formatting VendorPay project

+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Controllers/PayrollUserController.cs
@@ -1,27 +1,20 @@
-﻿using BTCPayServer.Abstractions.Constants;
+﻿using System;
+using System.Linq;
+using System.Security.Cryptography;
+using System.Threading.Tasks;
+using BTCPayServer.Abstractions.Constants;
 using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Abstractions.Models;
 using BTCPayServer.Client;
 using BTCPayServer.Data;
 using BTCPayServer.RockstarDev.Plugins.Payroll.Data;
 using BTCPayServer.RockstarDev.Plugins.Payroll.Data.Models;
+using BTCPayServer.RockstarDev.Plugins.Payroll.Services;
+using BTCPayServer.RockstarDev.Plugins.Payroll.Services.Helpers;
+using BTCPayServer.RockstarDev.Plugins.Payroll.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Text;
-using System.IO;
-using System.IO.Compression;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-using BTCPayServer.Abstractions.Contracts;
-using BTCPayServer.RockstarDev.Plugins.Payroll.ViewModels;
-using BTCPayServer.RockstarDev.Plugins.Payroll.Services;
-using System.Security.Cryptography;
-using BTCPayServer.RockstarDev.Plugins.Payroll.Services.Helpers;
-using BTCPayServer.Services.Mails;
 
 namespace BTCPayServer.RockstarDev.Plugins.Payroll.Controllers;
 
@@ -35,8 +28,8 @@ public class PayrollUserController(
     InvoicesDownloadHelper invoicesDownloadHelper)
     : Controller
 {
-    public StoreData CurrentStore => HttpContext.GetStoreData();
     private const string UserInviteEmailSubject = "You are invited to create a Vendor Pay account";
+
     private const string UserInviteEmailBody = @"Hello {Name},
 
 You are invited to create an account on {StoreName}'s Vendor Pay portal by visiting the following link:  
@@ -53,21 +46,18 @@ Payments will be issued in accordance with the terms of the contracted payment a
 Thank you,  
 {StoreName}";
 
+    public StoreData CurrentStore => HttpContext.GetStoreData();
+
 
     [HttpGet("list")]
     public async Task<IActionResult> List(string storeId, bool all, bool pending)
     {
         await using var ctx = pluginDbContextFactory.CreateContext();
-        IQueryable<PayrollUser> query = ctx.PayrollUsers.Where(a => a.StoreId == storeId);
+        var query = ctx.PayrollUsers.Where(a => a.StoreId == storeId);
         if (pending)
-        {
             query = query.Where(a => a.State == PayrollUserState.Pending);
-        }
-        else if (!all)
-        {
-            query = query.Where(a => a.State == PayrollUserState.Active);
-        }
-        List<PayrollUser> payrollUsers = query.OrderByDescending(data => data.Name).ToList();
+        else if (!all) query = query.Where(a => a.State == PayrollUserState.Active);
+        var payrollUsers = query.OrderByDescending(data => data.Name).ToList();
         var payrollUserListViewModel = new PayrollUserListViewModel
         {
             All = all,
@@ -85,9 +75,10 @@ Thank you,
             return NotFound();
 
         var isEmailSettingsConfigured = await emailService.IsEmailSettingsConfigured(CurrentStore.Id);
-        var vm = new PayrollUserCreateViewModel { 
-            StoreId = CurrentStore.Id, 
-            UserInviteEmailBody = UserInviteEmailBody, 
+        var vm = new PayrollUserCreateViewModel
+        {
+            StoreId = CurrentStore.Id,
+            UserInviteEmailBody = UserInviteEmailBody,
             UserInviteEmailSubject = UserInviteEmailSubject,
             StoreEmailSettingsConfigured = isEmailSettingsConfigured
         };
@@ -109,6 +100,7 @@ Thank you,
             ModelState.AddModelError(nameof(model.Email), "User with the same email already exists");
             return View(model);
         }
+
         var uid = Guid.NewGuid().ToString();
         var dbUser = new PayrollUser
         {
@@ -126,11 +118,11 @@ Thank you,
             {
                 TempData.SetStatusMessageModel(new StatusMessageModel
                 {
-                    Message = "An invitation has already been sent to this user",
-                    Severity = StatusMessageModel.StatusSeverity.Warning
+                    Message = "An invitation has already been sent to this user", Severity = StatusMessageModel.StatusSeverity.Warning
                 });
                 return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
             }
+
             var invitation = new PayrollInvitation
             {
                 Id = uid,
@@ -142,22 +134,22 @@ Thank you,
             };
             try
             {
-                await emailService.SendUserInvitationEmail(dbUser, model.UserInviteEmailSubject, model.UserInviteEmailBody, 
+                await emailService.SendUserInvitationEmail(dbUser, model.UserInviteEmailSubject, model.UserInviteEmailBody,
                     Url.Action("AcceptInvitation", "Public", new { storeId = CurrentStore.Id, invitation.Token }, Request.Scheme));
             }
             catch (Exception)
             {
-                ModelState.AddModelError(nameof(model.Email), 
+                ModelState.AddModelError(nameof(model.Email),
                     "Invitation email sending failed, kindly check your SMTP settings and ensure they are correct");
                 return View(model);
             }
+
             dbPlugins.Add(dbUser);
             dbPlugins.Add(invitation);
             await dbPlugins.SaveChangesAsync();
             TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = "Invitation email sent to "+ model.Email,
-                Severity = StatusMessageModel.StatusSeverity.Success
+                Message = "Invitation email sent to " + model.Email, Severity = StatusMessageModel.StatusSeverity.Success
             });
         }
         else
@@ -167,15 +159,16 @@ Thank you,
                 ModelState.AddModelError(nameof(model.Password), "Password cannot be empty");
                 return View(model);
             }
+
             dbUser.Password = hasher.HashPassword(uid, model.Password);
             dbPlugins.Add(dbUser);
             await dbPlugins.SaveChangesAsync();
             TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = "User created successfully",
-                Severity = StatusMessageModel.StatusSeverity.Success
+                Message = "User created successfully", Severity = StatusMessageModel.StatusSeverity.Success
             });
         }
+
         return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
     }
 
@@ -186,17 +179,16 @@ Thank you,
             return NotFound();
 
         await using var ctx = pluginDbContextFactory.CreateContext();
-        PayrollUser user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id && a.State == PayrollUserState.Pending);
+        var user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id && a.State == PayrollUserState.Pending);
         if (user == null)
         {
-            TempData.SetStatusMessageModel(new StatusMessageModel()
+            TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = $"Cannot send reminder to user. Invalid user state specified",
-                Severity = StatusMessageModel.StatusSeverity.Error
+                Message = "Cannot send reminder to user. Invalid user state specified", Severity = StatusMessageModel.StatusSeverity.Error
             });
             return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
-
         }
+
         var existingInvitation = ctx.PayrollInvitations.FirstOrDefault(i => i.Email == user.Email && i.StoreId == CurrentStore.Id);
         existingInvitation.Token = GenerateUniqueToken();
         existingInvitation.CreatedAt = DateTime.UtcNow;
@@ -207,19 +199,19 @@ Thank you,
         }
         catch (Exception)
         {
-            TempData.SetStatusMessageModel(new StatusMessageModel()
+            TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = $"Invitation email resending failed, kindly check your SMTP settings and ensure they are correct",
+                Message = "Invitation email resending failed, kindly check your SMTP settings and ensure they are correct",
                 Severity = StatusMessageModel.StatusSeverity.Error
             });
             return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
         }
+
         ctx.Update(existingInvitation);
         await ctx.SaveChangesAsync();
-        TempData.SetStatusMessageModel(new StatusMessageModel()
+        TempData.SetStatusMessageModel(new StatusMessageModel
         {
-            Message = $"User invitation successfully sent",
-            Severity = StatusMessageModel.StatusSeverity.Success
+            Message = "User invitation successfully sent", Severity = StatusMessageModel.StatusSeverity.Success
         });
         return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id, pending = true });
     }
@@ -239,7 +231,10 @@ Thank you,
 
         var model = new PayrollUserCreateViewModel
         {
-            Id = user.Id, Email = user.Email, Name = user.Name, EmailReminder = user.EmailReminder, 
+            Id = user.Id,
+            Email = user.Email,
+            Name = user.Name,
+            EmailReminder = user.EmailReminder,
             StoreEmailSettingsConfigured = await emailService.IsEmailSettingsConfigured(CurrentStore.Id)
         };
         return View(model);
@@ -259,13 +254,14 @@ Thank you,
 
         await using var ctx = pluginDbContextFactory.CreateContext();
 
-        PayrollUser user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
+        var user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
 
         user.Email = string.IsNullOrEmpty(model.Email) ? user.Email : model.Email;
         user.Name = string.IsNullOrEmpty(model.Name) ? user.Name : model.Name;
 
-        user.EmailReminder = string.IsNullOrEmpty(model.EmailReminder) ? user.EmailReminder : 
-            string.Join(",", model.EmailReminder.Split(',')
+        user.EmailReminder = string.IsNullOrEmpty(model.EmailReminder)
+            ? user.EmailReminder
+            : string.Join(",", model.EmailReminder.Split(',')
                 .Select(r => r.Trim()).Where(r => !string.IsNullOrEmpty(r))
                 .Distinct().OrderBy(int.Parse));
 
@@ -285,10 +281,10 @@ Thank you,
 
         await using var ctx = pluginDbContextFactory.CreateContext();
 
-        PayrollUser user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
+        var user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
         if (user.State == PayrollUserState.Pending)
             return NotFound();
-        PayrollUserResetPasswordViewModel model = new PayrollUserResetPasswordViewModel { Id = user.Id };
+        var model = new PayrollUserResetPasswordViewModel { Id = user.Id };
         return View(model);
     }
 
@@ -300,7 +296,7 @@ Thank you,
 
         await using var ctx = pluginDbContextFactory.CreateContext();
 
-        PayrollUser user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
+        var user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
 
         if (!ModelState.IsValid)
             return View(model);
@@ -322,12 +318,14 @@ Thank you,
             return NotFound();
 
         await using var ctx = pluginDbContextFactory.CreateContext();
-        PayrollUser user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
+        var user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
 
         if (user == null)
             return NotFound();
 
-        return View("Confirm", new ConfirmModel($"{(enable ? "Activate" : "Disable")} user", $"The user ({user.Name}) will be {(enable ? "activated" : "disabled")}. Are you sure?", (enable ? "Activate" : "Disable")));
+        return View("Confirm",
+            new ConfirmModel($"{(enable ? "Activate" : "Disable")} user",
+                $"The user ({user.Name}) will be {(enable ? "activated" : "disabled")}. Are you sure?", enable ? "Activate" : "Disable"));
     }
 
 
@@ -338,7 +336,7 @@ Thank you,
             return NotFound();
 
         await using var ctx = pluginDbContextFactory.CreateContext();
-        PayrollUser user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
+        var user = ctx.PayrollUsers.SingleOrDefault(a => a.Id == userId && a.StoreId == CurrentStore.Id);
 
         if (user == null)
             return NotFound();
@@ -355,6 +353,7 @@ Thank you,
                 // Would need to know use case for this.
                 break;
         }
+
         await ctx.SaveChangesAsync();
 
         TempData[WellKnownTempData.SuccessMessage] = $"User {(enable ? "activated" : "disabled")} successfully";
@@ -377,10 +376,9 @@ Thank you,
             .OrderByDescending(data => data.CreatedAt).ToListAsync();
         if (!payrollInvoices.Any())
         {
-            TempData.SetStatusMessageModel(new StatusMessageModel()
+            TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = $"No payroll invoice available for download",
-                Severity = StatusMessageModel.StatusSeverity.Info
+                Message = "No payroll invoice available for download", Severity = StatusMessageModel.StatusSeverity.Info
             });
             return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
         }
@@ -391,10 +389,9 @@ Thank you,
 
     private void ReturnMessageStatus()
     {
-        TempData.SetStatusMessageModel(new StatusMessageModel()
+        TempData.SetStatusMessageModel(new StatusMessageModel
         {
-            Message = $"User details updated successfully",
-            Severity = StatusMessageModel.StatusSeverity.Success
+            Message = "User details updated successfully", Severity = StatusMessageModel.StatusSeverity.Success
         });
     }
 
@@ -411,23 +408,23 @@ Thank you,
             return NotFound();
 
         if (ctx.PayrollInvoices.Any(a => a.UserId == user.Id &&
-            (a.State != PayrollInvoiceState.Completed && a.State != PayrollInvoiceState.Cancelled)))
+                                         a.State != PayrollInvoiceState.Completed && a.State != PayrollInvoiceState.Cancelled))
         {
-            TempData.SetStatusMessageModel(new StatusMessageModel()
+            TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = $"User can't be deleted since there are active invoices for this user",
-                Severity = StatusMessageModel.StatusSeverity.Error
+                Message = "User can't be deleted since there are active invoices for this user", Severity = StatusMessageModel.StatusSeverity.Error
             });
             return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
         }
-        var completedOrCancelledInvoices = ctx.PayrollInvoices.Count(a => 
+
+        var completedOrCancelledInvoices = ctx.PayrollInvoices.Count(a =>
             a.UserId == user.Id && (a.State == PayrollInvoiceState.Completed || a.State == PayrollInvoiceState.Cancelled));
 
-        string invoiceDeleteText = completedOrCancelledInvoices > 0
-            ? $"The user: {user.Name} will be deleted along with {completedOrCancelledInvoices} associated invoices. Are you sure you want to proceed?" : 
-            $"The user: {user.Name} will be deleted. Are you sure?";
+        var invoiceDeleteText = completedOrCancelledInvoices > 0
+            ? $"The user: {user.Name} will be deleted along with {completedOrCancelledInvoices} associated invoices. Are you sure you want to proceed?"
+            : $"The user: {user.Name} will be deleted. Are you sure?";
 
-        return View("Confirm", new ConfirmModel($"Delete user", invoiceDeleteText, "Delete"));
+        return View("Confirm", new ConfirmModel("Delete user", invoiceDeleteText, "Delete"));
     }
 
 
@@ -444,33 +441,26 @@ Thank you,
         var userInvoices = ctx.PayrollInvoices.Where(a => a.UserId == payrollUser.Id).ToList();
         if (userInvoices.Any(a => a.State != PayrollInvoiceState.Completed && a.State != PayrollInvoiceState.Cancelled))
         {
-            TempData.SetStatusMessageModel(new StatusMessageModel()
+            TempData.SetStatusMessageModel(new StatusMessageModel
             {
-                Message = $"User can't be deleted since there are active invoices for this user",
-                Severity = StatusMessageModel.StatusSeverity.Error
+                Message = "User can't be deleted since there are active invoices for this user", Severity = StatusMessageModel.StatusSeverity.Error
             });
             return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
         }
+
         var payrollUserInvite = ctx.PayrollInvitations.Where(c => c.Email == payrollUser.Email && c.StoreId == payrollUser.StoreId).ToList();
-        if (payrollUserInvite.Any())
-        {
-            ctx.RemoveRange(payrollUserInvite);
-        }
+        if (payrollUserInvite.Any()) ctx.RemoveRange(payrollUserInvite);
         ctx.RemoveRange(userInvoices);
         ctx.Remove(payrollUser);
         await ctx.SaveChangesAsync();
 
-        TempData.SetStatusMessageModel(new StatusMessageModel()
-        {
-            Message = $"User deleted successfully",
-            Severity = StatusMessageModel.StatusSeverity.Success
-        });
+        TempData.SetStatusMessageModel(new StatusMessageModel { Message = "User deleted successfully", Severity = StatusMessageModel.StatusSeverity.Success });
         return RedirectToAction(nameof(List), new { storeId = CurrentStore.Id });
     }
 
     public string GenerateUniqueToken()
     {
-        byte[] tokenData = new byte[32];
+        var tokenData = new byte[32];
         RandomNumberGenerator.Fill(tokenData);
         return Convert.ToBase64String(tokenData)
             .Replace("+", "-")
